+++ doc/source/changelog.rst
@@ -29,6 +29,8 @@ organisation on `GitHub <https://github.com/openbiosim/sire>`__.
 
 * Null SystemTrajectory pointer when all frames are deleted.
 
+* Handle cis/trans double bond stereochemistry values in SDF bond blocks.
+
 * Fix handling of positive formal charge when writing SDF files.
 
 `2024.3.1 <https://github.com/openbiosim/sire/compare/2024.3.0...2024.3.1>`__ - December 2024